move encrypted note code to its own file
